If Opera is used, get the xlink:href attribute with namespace (fixes MAP-484).
This would also work in FireFox, but not in IE and actually applies the idea of r3024 by jseb.baklouti to all possible places.

For OPera MapPaneOL.js makes sure that the height of the map is > 0, and optionally calculates the height out of a %-value, with a fallback. This was a problem at the gmlviewer, ol-integration and i18n examples