Add support for IE10: handle IE10 like IE9
OR
maybe also opt-in to retrieving an MSXML document. This is  useful if you still need some MSXML-specific functionality (such as selectNodes). To do this, set the responseType of your XMLHttpRequest object to 'msxml-document'.
That would be an extra line 
try { xmlhttp.responseType = 'msxml-document'; } catch(e){}
in ModelBase.js
- see http://blogs.msdn.com/b/ie/archive/2012/07/19/xmlhttprequest-responsexml-in-ie10-release-preview.aspx


examples: use correct namespace for variable to decide, if the browser is an IE